+++ fiberplane-models/src/webhooks.rs
@@ -86,6 +86,10 @@ pub struct Webhook {
     #[builder(default)]
     pub enabled: bool,
 
+    /// Whether the last delivery was successful
+    #[builder(default)]
+    pub successful: bool,
+
     #[builder(default, setter(strip_option, into))]
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub created_by: Option<Base64Uuid>,
@@ -106,7 +110,11 @@ pub struct Webhook {
 pub struct NewWebhook {
     #[builder(setter(into))]
     pub endpoint: String,
+
     pub events: Vec<WebhookCategory>,
+
+    #[builder(default)]
+    pub enabled: bool,
 }
 
 #[derive(Clone, Debug, Deserialize, PartialEq, Eq, Serialize, TypedBuilder)]